+++ src/main.rs
@@ -59,7 +59,7 @@ pub fn run_command(ctx: &Context, command: Command) -> Result<()> {
 ///
 /// Initialize a new config file.
 fn init(ctx: &Context, shell: Option<Shell>) -> Result<()> {
-    let _guard = access(ctx, Access::W)?;
+    let _guard = access_ignore_not_found(ctx, Access::W)?;
     let path = ctx.config_file();
     match path
         .metadata()
@@ -80,7 +80,7 @@ fn init(ctx: &Context, shell: Option<Shell>) -> Result<()> {
 ///
 /// Add a new plugin to the config file.
 fn add(ctx: &Context, name: String, plugin: &EditPlugin) -> Result<()> {
-    let _guard = access(ctx, Access::W)?;
+    let _guard = access_ignore_not_found(ctx, Access::W)?;
     let path = ctx.config_file();
     let mut config = match EditConfig::from_path(path) {
         Ok(config) => {
@@ -100,7 +100,7 @@ fn add(ctx: &Context, name: String, plugin: &EditPlugin) -> Result<()> {
 ///
 /// Open up the config file in the default editor.
 fn edit(ctx: &Context) -> Result<()> {
-    let _guard = access(ctx, Access::W)?;
+    let _guard = access_ignore_not_found(ctx, Access::W)?;
     let path = ctx.config_file();
     let original_contents = match fs::read_to_string(path)
         .with_context(|| format!("failed to read from `{}`", path.display()))
@@ -129,7 +129,7 @@ fn edit(ctx: &Context) -> Result<()> {
 ///
 /// Remove a plugin from the config file.
 fn remove(ctx: &Context, name: String) -> Result<()> {
-    let _guard = access(ctx, Access::W)?;
+    let _guard = access_ignore_not_found(ctx, Access::W)?;
     let path = ctx.config_file();
     let mut config = EditConfig::from_path(path)?;
     ctx.log_header("Loaded", path);
@@ -269,10 +269,18 @@ enum Access {
     W,
 }
 
+fn access_ignore_not_found(ctx: &Context, mode: Access) -> Result<Option<fmutex::Guard<'static>>> {
+    match access(ctx, mode) {
+        Ok(guard) => Ok(Some(guard)),
+        Err(err) if underlying_io_error_kind(&err) == Some(io::ErrorKind::NotFound) => Ok(None),
+        Err(err) => Err(err),
+    }
+}
+
 fn access(ctx: &Context, mode: Access) -> Result<fmutex::Guard<'static>> {
     match mode {
-        Access::R => lock_read(ctx).context("failed to acquire exclusive lock on config directory"),
-        Access::W => lock_write(ctx).context("failed to acquire shared lock on config directory"),
+        Access::R => lock_read(ctx).context("failed to acquire shared lock on config directory"),
+        Access::W => lock_write(ctx).context("failed to acquire lock on config directory"),
     }
 }